undo manual changes
